correct swagger code gen version
