+++ slack-message-reader-api/src/main/java/cd/connect/samples/slackapp/api/jaxrs/JaxrsClient.java
@@ -1,0 +1,7 @@
+package cd.connect.samples.slackapp.api.jaxrs;
+
+import javax.ws.rs.client.Client;
+
+public interface JaxrsClient {
+  Client getClient();
+}
