--- slack-message-reader-api/src/main/java/co/connect/samples/slackapp/api/jaxrs/JaxrsClient.java
@@ -1,7 +1,0 @@
-package co.connect.samples.slackapp.api.jaxrs;
-
-import javax.ws.rs.client.Client;
-
-public interface JaxrsClient {
-  Client getClient();
-}
