--- slack-message-reader-api/src/main/resources/api.yaml
@@ -1,58 +1,0 @@
-swagger: '2.0'
-info:
-  version: ''
-  title: SlackMessages
-  description: Slack Message Query API
-host: 'localhost:8080'
-basePath: /slack-messages-api
-schemes:
-  - http
-paths:
-  /messages:
-    get:
-      operationId: GET_messages
-      summary: get messages
-      consumes:
-        - application/json
-      produces:
-        - application/json
-      parameters:
-        - name: body
-          in: body
-          schema:
-            $ref: '#/definitions/messagesquery'
-      responses:
-        '200':
-          description: ''
-          schema:
-            $ref: '#/definitions/messagelist'
-          examples:
-            application/json:
-              - {}
-              - {}
-definitions:
-  messagesquery:
-    title: messageQuery
-    type: object
-    properties:
-      userId:
-        type: string
-      fromDate:
-        type: string
-      toDate:
-        type: string
-  messagelist:
-    title: messageList
-    type: array
-    items:
-      $ref: '#/definitions/message'
-  message:
-    title: message
-    type: object
-    properties:
-      userId:
-        type: string
-      content:
-        type: string
-      timestamp:
-        type: number
