+++ slack-message-reader-api/src/main/resources/generator/jersey-proxy-module.mustache
@@ -1,7 +1,7 @@
 package {{packageName}};
 
-import co.connect.samples.slackapp.api.config.ConfigConst;
-import co.connect.samples.slackapp.api.jaxrs.JaxrsClientModule;
+import cd.connect.samples.slackapp.api.config.ConfigConst;
+import cd.connect.samples.slackapp.api.jaxrs.JaxrsClientModule;
 
 {{#jersey}}
   // plain resources
