+++ slack-message-reader/pom.xml
@@ -15,6 +15,11 @@
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
+            <groupId>cd.connect.composites.java</groupId>
+            <artifactId>connect-composite-log4j2</artifactId>
+            <version>1.2-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>4.12</version>
@@ -35,10 +40,21 @@
             <version>1.17</version>
         </dependency>
         <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-webmvc</artifactId>
-            <version>4.3.9.RELEASE</version>
+            <groupId>io.prometheus</groupId>
+            <artifactId>simpleclient_servlet</artifactId>
+            <version>0.0.23</version>
         </dependency>
+        <dependency>
+            <groupId>commons-logging</groupId>
+            <artifactId>commons-logging</artifactId>
+            <version>1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>cd.connect.servlet</groupId>
+            <artifactId>servlet-spring-jersey</artifactId>
+            <version>1.5-SNAPSHOT</version>
+        </dependency>
+
     </dependencies>
 
     <build>
