+++ slack-message-reader/src/main/java/cd/connect/samples/slackapp/Application.java
@@ -1,0 +1,12 @@
+package cd.connect.samples.slackapp;
+
+import cd.connect.samples.slackapp.api.SlackMessageyApiModule;
+import cd.connect.samples.slackapp.api.jaxrs.JaxrsClientManager;
+import cd.connect.samples.slackapp.api.jaxrs.JaxrsClientModule;
+import org.springframework.context.annotation.Import;
+
+@Import({SlackAppGenConfig.class, JerseyDataConfig.class, SlackMessageyApiModule.class})
+public class Application extends BaseApplication {
+
+
+}
