+++ slack-message-reader/src/main/java/cd/connect/samples/slackapp/dao/MessageDao.java
@@ -1,0 +1,4 @@
+package cd.connect.samples.slackapp.dao;
+
+public class MessageDao {
+}
