--- slack-message-reader/src/main/java/co/connect/samples/slackapp/Appliacation.java
@@ -1,10 +1,0 @@
-package co.connect.samples.slackapp;
-
-import co.connect.samples.slackapp.api.SlackMessageyApiModule;
-import org.springframework.context.annotation.Import;
-
-@Import({SlackMessageyApiModule.class})
-public class Appliacation extends BaseApplication {
-
-
-}
