--- slack-message-reader/src/main/java/co/connect/samples/slackapp/dao/MessageDao.java
@@ -1,4 +1,0 @@
-package co.connect.samples.slackapp.dao;
-
-public class MessageDao {
-}
