+++ slack-message-reader/src/main/resources/generator/common-spring.mustache
@@ -1,0 +1,47 @@
+package {{packageName}};
+
+import java.util.stream.Stream;
+import cd.connect.spring.servlet.ServletModule;
+{{#spring}}
+  // plain resources
+  {{#sortedTypes}}
+import {{packageName}}.{{name}};
+  {{/sortedTypes}}
+{{/spring}}
+{{#servlet}}
+  // servlets that have WebServlet annotations
+  {{#sortedTypes}}
+import {{packageName}}.{{name}};
+  {{/sortedTypes}}
+{{/servlet}}
+{{#filter}}
+  // filters that have WebFilter annotations
+  {{#sortedTypes}}
+import {{packageName}}.{{name}};
+  {{/sortedTypes}}
+{{/filter}}
+
+public class {{simpleName}} extends ServletModule {
+  public void register() {
+  {{#spring}}
+    register(Stream.of(
+    {{#sortedTypes}}
+      {{name}}.class{{^-last}}, {{/-last}}
+    {{/sortedTypes}}
+    ));
+  {{/spring}}
+{{#servlet}}
+  // servlets that have WebServlet annotations
+  {{#sortedTypes}}
+    servlet({{name}}.class);
+  {{/sortedTypes}}
+{{/servlet}}
+{{#filter}}
+  // filters that have WebFilter annotations
+  {{#sortedTypes}}
+    filter({{name}}.class);
+  {{/sortedTypes}}
+{{/filter}}
+
+  }
+}
