+++ slack-message-reader/src/main/resources/generator/jersey.mustache
@@ -1,0 +1,43 @@
+package {{packageName}};
+
+import java.util.stream.Stream;
+import cd.connect.samples.slackapp.servlet.JerseyModule;
+{{#jersey}}
+  // plain resources
+  {{#sortedTypes}}
+import {{packageName}}.{{name}};
+  {{/sortedTypes}}
+{{/jersey}}
+
+public class {{simpleName}} extends JerseyModule {
+  // register these with Spring
+  public void register() {
+{{#jersey}}
+    register(Stream.of(
+  {{#sortedTypes}}
+    {{name}}.class{{^-last}}, {{/-last}}
+  {{/sortedTypes}}
+  ));
+{{/jersey}}
+  }
+
+  // tell the JerseyApplication to extract these from the Spring Context and register them
+  // with this Jersey Application
+  protected Stream<Class<?>> registerResources() {
+{{#jersey}}
+    return Stream.of(
+  {{#sortedTypes}}
+    {{name}}.class{{^-last}}, {{/-last}}
+  {{/sortedTypes}});
+{{/jersey}}
+  }
+
+  // url for no cache f
+  protected String getUrlNoCachePattern() {
+    return "{{{urlPattern}}}";
+  }
+
+  protected String getUrlServletPrefix() {
+    return "{{{baseUrl}}}";
+  }
+}
