+++ slack-message-reader/src/test/resources/war.properties
@@ -1,0 +1,1 @@
+webapp.http.port=8090
